+++ packages/media/src/audio/sort-by-priority.ts
@@ -1,6 +1,5 @@
 type Waiter = {
-	getPriority: () => number;
-	getStale: () => boolean;
+	getPriority: () => number | null;
 	fn: () => Promise<unknown>;
 	onDone: (result: unknown, triggerNext: () => void) => void;
 	onError: (err: unknown) => void;
@@ -29,7 +28,7 @@ const processNext = (): void => {
 	// the iteration and leave `waiters[i]` undefined on the next `i--`.
 	const staleWaiters: Waiter[] = [];
 	for (let i = waiters.length - 1; i >= 0; i--) {
-		if (waiters[i].getStale()) {
+		if (waiters[i].getPriority() === null) {
 			const [stale] = waiters.splice(i, 1);
 			staleWaiters.push(stale);
 		}
@@ -45,8 +44,16 @@ const processNext = (): void => {
 
 	let bestIndex = 0;
 	let bestPriority = waiters[0].getPriority();
+	if (bestPriority === null) {
+		throw new Error('Stale waiter should have been removed');
+	}
+
 	for (let i = 1; i < waiters.length; i++) {
 		const priority = waiters[i].getPriority();
+		if (priority === null) {
+			throw new Error('Stale waiter should have been removed');
+		}
+
 		if (priority < bestPriority) {
 			bestPriority = priority;
 			bestIndex = i;
@@ -70,20 +77,17 @@ const processNext = (): void => {
 
 export const waitForTurn = <T>({
 	getPriority,
-	getStale,
 	fn,
 	onDone,
 	onError,
 }: {
-	getPriority: () => number;
-	getStale: () => boolean;
+	getPriority: () => number | null;
 	fn: () => Promise<T>;
 	onDone: (result: T, triggerNext: () => void) => void;
 	onError: (err: unknown) => void;
 }): void => {
 	waiters.push({
 		getPriority,
-		getStale,
 		fn,
 		onDone: onDone as (result: unknown, triggerNext: () => void) => void,
 		onError: onError as (err: unknown) => void,
