+++ packages/media/src/test/audio-cleanup-on-seek.test.ts
@@ -53,7 +53,6 @@ const makeMockSharedAudioContext = ({
 		} as unknown as AudioContext,
 		audioSyncAnchor: {value: anchorValue},
 		scheduleAudioNode: () => ({type: 'started', scheduledTime: 0}),
-		getScheduledTime: () => 0,
 		getDurationOfNode: () => 0,
 	};
 };
